+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 修改 `Asset/RemoteUri` 的默认值为 `Builds/Patch/${Env.Author}/${Env.Version}/${Env.Platform}/Assets`
 - 修改 `Const.Manifest` 的默认值为 `Assets`
 - 新增 Build Assets 和 Publish Assets 任务配置项的 GUI 面板显示
+- 移除 XAsset.Utility 的 Loading 函数并重构为 XAsset.Resource/Scene.IsLoading
 
 ## [0.0.8] - 2025-06-17
 ### 变更